Add files via upload
The results of this challenge for me were quite interesting, in terms of how illiterate and messy it was. For some reason my dice kept on landing on the option where I had to either draw with my mouth or elbow. The results were very much distorted but kind of interesting. I guess you could say this is what happens when you try to input an algorithm that is either incorrect or too much for the computer to handle so therefore it lags and creates a jumble of messy lines.  

And, don’t even get me started on drawing with your knees option as that was just impossible! This was not due to my lack of trying, because I tried and failed multiple times. 

Also, if anyone is wondering what the bottom left corner was meant to be. They were supposed to be hands, but the dice landed twice on drawing hands with my mouth.